autocompletion

+++ computer-example/BOOTLOADER/backupHandler.lua
@@ -26,6 +26,11 @@ function Backup_Handler.cleanup(path)
 end
 
 function Backup_Handler.restoreBackup(path) -- do not include the "-BACKUP" in path
+
+    if not fs.exists(path.."-BACKUP") then
+        return false, "No backup for this file"
+    end
+
     local restoreBackup_success, restoreBackup_error = pcall(function()
         fs.delete(path)
         fs.copy(path.."-BACKUP",path)
